# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: TEENet Wallet API
  description: |
    Multi-chain crypto wallet backed by TEE-DAO threshold cryptography.
    Private keys never leave TEE hardware -- they are split across distributed
    TEE-DAO nodes via threshold signing (FROST for Ed25519, GG20 for ECDSA).

    ## Alpha Release

    The public TEENet Wallet is currently in alpha:

    - **Testnet only (gated by `ALPHA_MODE`).** The public deployment runs
      with `ALPHA_MODE=true`, which at startup drops every chain whose
      `testnet` field is not `true` from the registry. `GET /api/chains`
      therefore returns 8 testnets (`sepolia`, `optimism-sepolia`,
      `arbitrum-sepolia`, `base-sepolia`, `polygon-amoy`, `bsc-testnet`,
      `avalanche-fuji`, `solana-devnet`). Mainnet chains (`ethereum`,
      `optimism`, `arbitrum`, `base`, `polygon`, `bsc`, `avalanche`, `solana`)
      are shipped in `chains.json` and fully supported by the API — they are
      just filtered out during alpha. Clearing `ALPHA_MODE` re-enables them.
    - **Registration cap.** Public registration is limited to the first 500
      users on a first-come, first-served basis (`MAX_USERS`, default `500`).
      Once the cap is reached, `POST /api/auth/passkey/register/verify` and
      the passkey auto-create path return HTTP 403 with
      `"maximum number of users reached"`.
    - **Breaking changes expected.** API shapes, error fields, and storage
      layouts may change before general availability.

...

...

...

...

...

...

...

...

...

...
servers:
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "API Key (ocw_ prefix) or Passkey session token (ps_ prefix)"
    PasskeySession:
      type: http
      scheme: bearer
      description: "Passkey session token (ps_ prefix) only"
  parameters:
    WalletId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Wallet UUID
      example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
    ContractId:
      name: cid
      in: path
      required: true
      schema:
        type: integer
      description: Allowed contract record ID
      example: 42
    ChainName:
      name: chain
      in: path
      required: true
      schema:
        type: string
      description: Chain name (e.g. `sepolia`, `base`, `solana-devnet`).
      example: "sepolia"
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      schema:
        type: string
      description: Required for state-changing requests from Passkey sessions
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Idempotency key to prevent duplicate transfers
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "wallet not found"
    SuccessFlag:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    AllowedContract:
      type: object
      properties:
        id:
          type: integer
          example: 1
        wallet_id:
          type: string
          format: uuid
        contract_address:
          type: string
          example: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48"
        label:
          type: string
          example: "USDC on Ethereum"
        symbol:
          type: string
          example: "USDC"
        decimals:
          type: integer
          example: 6
        created_at:
          type: string
          format: date-time
...
    TokenParams:
      type: object
      required: [contract]
      properties:
        contract:
          type: string
          description: ERC-20 contract address or SPL token mint
          example: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48"
        decimals:
          type: integer
          description: Token decimals (default 18 for EVM, 9 for Solana)
          example: 6
        symbol:
          type: string
          example: "USDC"
...
    PendingApprovalResponse:
      type: object
      properties:
        status:
          type: string
          example: "pending_approval"
        approval_id:
          type: integer
          example: 4839271056
        message:
          type: string
          example: "amount ~$150.00 USD exceeds approval threshold $100 USD"
        tx_context:
          type: object
        approval_url:
          type: string
          example: "http://localhost:18080/#/approve/4839271056"
...
    FreshPasskeyCredential:
      type: object
      description: WebAuthn assertion for fresh passkey verification
      required: [login_session_id, credential]
      properties:
        login_session_id:
          type: integer
          format: uint64
          example: 12345
        credential:
          type: object
          description: WebAuthn credential assertion response
paths:
...
  /api/wallets/{id}/transfer:
    post:
      tags: [Wallets]
      summary: Build, sign, and broadcast a transfer
      description: |
        Builds a blockchain transaction server-side, signs it via TEE threshold
        signing, and broadcasts it. Supports native tokens (ETH, SOL) and
        token transfers (ERC-20, SPL) when a `token` object is provided.

        If an approval policy is configured and the USD amount exceeds the
        threshold or daily limit, the request returns 202 with a pending approval.

        Supports `Idempotency-Key` header to prevent duplicate transactions.
      operationId: transfer
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/WalletId"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, amount]
              properties:
                to:
                  type: string
                  description: Recipient address (0x... for EVM, base58 for Solana)
                  example: "0xRecipientAddress..."
                amount:
                  type: string
                  description: Amount in human-readable units (e.g. "1.5" ETH, "0.01" SOL)
                  example: "0.1"
                memo:
                  type: string
                  description: Optional memo/note
                  example: "Payment for services"
                token:
                  $ref: "#/components/schemas/TokenParams"
      responses:
        "200":
          description: Transfer completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "completed"
                  tx_hash:
                    type: string
                    example: "0xabc123..."
                  chain:
                    type: string
                    example: "sepolia"
                  from:
                    type: string
                    example: "0xSenderAddress..."
                  to:
                    type: string
                    example: "0xRecipientAddress..."
                  amount:
                    type: string
                    example: "0.1"
                  currency:
                    type: string
                    example: "ETH"
        "202":
          description: Pending approval (USD threshold exceeded)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingApprovalResponse"
        "400":
          description: Validation error, wallet not ready, or daily limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Token contract not whitelisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: RPC, signing, or broadcast failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/chains/{chain}/contracts:
    get:
      tags: [Contracts]
      summary: List whitelisted contracts on a chain
      operationId: listContracts
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/ChainName"
      responses:
        "200":
          description: Contract list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  contracts:
                    type: array
                    items:
                      $ref: "#/components/schemas/AllowedContract"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [Contracts]
      summary: Add a contract to the chain's whitelist
      description: |
        Whitelists a contract address on a chain. No wallet required —
        the whitelist is keyed per (user, chain).
        - **Passkey session**: applied immediately (requires fresh credential).
        - **API key**: creates a pending approval for the Passkey owner to review.
      operationId: addContract
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/ChainName"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contract_address]
              properties:
                login_session_id:
                  type: integer
                  format: uint64
                  description: Required for Passkey sessions
                credential:
                  type: object
                  description: Fresh WebAuthn assertion (required for Passkey sessions)
                contract_address:
                  type: string
                  description: EVM contract address (0x...) or Solana program ID (base58)
                  example: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48"
                label:
                  type: string
                  example: "USDC on Ethereum"
                symbol:
                  type: string
                  example: "USDC"
                decimals:
                  type: integer
                  minimum: 0
                  maximum: 77
                  example: 6
      responses:
        "200":
          description: Contract added (Passkey path)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  contract:
                    $ref: "#/components/schemas/AllowedContract"
        "202":
          description: Pending approval (API key path)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  pending:
                    type: boolean
                    example: true
                  approval_id:
                    type: integer
                    example: 4839271056
                  message:
                    type: string
                    example: "Contract whitelist request submitted for approval"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated or passkey verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Contract already whitelisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/chains/{chain}/contracts/{cid}:
    put:
      tags: [Contracts]
      summary: Rename a whitelisted contract
      description: |
        Updates the display label of an existing whitelisted contract.
        Only `label` is editable: `symbol` and `decimals` are on-chain
        metadata captured at whitelist time and are immutable here
        (re-add the contract if they need to change).

        Applied immediately for both passkey sessions and API-key
        requests — no approval is required because the label is
        display-only and has no effect on transfer semantics or which
        contracts are whitelisted.
      operationId: updateContract
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/ChainName"
        - $ref: "#/components/parameters/ContractId"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label:
                  type: string
                  description: New display label for the contract.
      responses:
        "200":
          description: Label updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  contract:
                    $ref: "#/components/schemas/AllowedContract"
        "400":
          description: Validation error (label missing or bad JSON)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Contracts]
      summary: Remove a whitelisted contract
      description: Removes a contract from the whitelist. Requires fresh passkey verification.
      operationId: deleteContract
      security:
        - PasskeySession: []
      parameters:
        - $ref: "#/components/parameters/ChainName"
        - $ref: "#/components/parameters/ContractId"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreshPasskeyCredential"
      responses:
        "200":
          description: Contract removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessFlag"
        "400":
          description: Missing credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Passkey verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not a Passkey session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Contract not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ──────────────────────── Contract Calls ────────────────────────
...